adds role based mixin
