+++ src/sentry/integrations/slack/message_builder/issues.py
@@ -27,8 +27,6 @@ from sentry.utils import json
 from sentry.utils.dates import to_timestamp
 from sentry.utils.http import absolute_uri
 
-from ..utils import build_notification_footer
-
 STATUSES = {"resolved": "resolved", "ignored": "ignored", "unresolved": "re-opened"}
 
 
@@ -342,7 +340,7 @@ class SlackIssuesMessageBuilder(SlackMessageBuilder):
         color = get_color(event_for_tags, self.notification)
         fields = build_tag_fields(event_for_tags, self.tags)
         footer = (
-            build_notification_footer(self.notification, self.recipient)
+            self.notification.build_notification_footer(self.recipient)
             if self.notification and self.recipient
             else build_footer(self.group, project, self.rules)
         )
