+++ src/sentry/notifications/notifications/activity/base.py
@@ -10,6 +10,7 @@ from django.utils.safestring import SafeString, mark_safe
 
 from sentry.notifications.helpers import get_reason_context
 from sentry.notifications.notifications.base import ProjectNotification
+from sentry.notifications.types import NotificationSettingTypes
 from sentry.notifications.utils import send_activity_notification
 from sentry.notifications.utils.avatar import avatar_as_html
 from sentry.notifications.utils.participants import get_participants_for_group
@@ -20,7 +21,7 @@ if TYPE_CHECKING:
 
 
 class ActivityNotification(ProjectNotification, ABC):
-    fine_tuning_key = "workflow"
+    notification_setting_type = NotificationSettingTypes.WORKFLOW
     metrics_key = "activity"
 
     def __init__(self, activity: Activity) -> None:
