+++ src/sentry/notifications/notifications/rules.py
@@ -7,7 +7,7 @@ import pytz
 
 from sentry.models import Team, User, UserOption
 from sentry.notifications.notifications.base import ProjectNotification
-from sentry.notifications.types import ActionTargetType
+from sentry.notifications.types import ActionTargetType, NotificationSettingTypes
 from sentry.notifications.utils import (
     get_commits,
     get_integration_link,
@@ -27,7 +27,7 @@ logger = logging.getLogger(__name__)
 
 class AlertRuleNotification(ProjectNotification):
     message_builder = "IssueNotificationMessageBuilder"
-    fine_tuning_key = "alerts"
+    notification_setting_type = NotificationSettingTypes.ISSUE_ALERTS
     metrics_key = "issue_alert"
 
     def __init__(
